Updating MsRejectWorker and adding function.
MsRejectWorker class is up and running. Also, added the
`assemble_msreject_workers` to create a list of MsRejectWorker
instances from simulated observed data. Some other miscellaneous
updates to workers module. The `parse_header` now checks if the
file has a header and raises an exception if it does not.
The Worker base class no longer has a TemporaryFileSystem as an
attribute.