Adding RegressionWorker class.
This class runs the post-rejection regression adjustment via
a subprocess that executes `regress_cli.r`. An instance of this
class can now also be passed to MsRejectWorkers on init, in which
case, the instance is run after the rejection subprocess finishes.
This has also been incorporated into the `assemble_msreject_workers`
function.

All of these changes still need to be tested.