Fixing bugs in worker classes.
The Worker base class was calling some old attributes in the
start method if there was an error during execution. All the
attributes are now updated. Also, adding a call to new
method `_pre_processing` in beginning of `start` method of the
Worker base class. This allows workers to defer some init stuff
until just before execution. This allows the rejection and regression
workers to be co-initiated (the RegressionWorkers need info in
the output files of the MsRejectWorkers).