Fixing bug in `merge_priors` function.
This function did not work correctly if the prior files had header
lines (i.e., they were getting dumped into the merged file). Now there
is a regex check to avoid this problem.

Also, I am changing the `observed` attribute of the MsBayesWorker
class to `include_header` (also updating the keyword for this
argument in init to include_header). This is more transparent.