Updating `get_stat_indices` function.
Removing odd pseudo-default option. The default pattern argument
are the `DEFAULT_STAT_PATTERNS`. But there was also an
`if not stat_patterns` clause in the function that set the patterns
to `ALL_STAT_PATTERNS`. This behavior is odd, and I do not know
why it's there, and it doesn't seem to be used anywhere in the code,
and so I am getting rid of it.