Incorporating staging directory for MsBayesWorker.
Adding an option to have a staging directory for MsBayesWorker
instances. This directory will be used to write the prior during
the msbayes.pl subprocess, and to convert the file format during
post-processing. Then, the final prior file is moved to the
directory within the temp file system. This allows the user
to provide a local, temporary staging directory on a compute node
to speed up the I/O during the subprocesses.