More overhauling of workers module and more tests.
More work on `ABCToolBoxRejectWorker` and `ABCToolBoxRegressWorker`
classes and on the `assemble_rejection_workers` method.

Having `ABCToolBoxRejectWorker` instance remove the extra columns
it creates in post-processing. This simplifies the assemble function,
because it doesn't have to account for the extra columns. Also,
I am having `ABCToolBoxRejectWorker` instances preserve 'dummy'
columns, which, again, simplifies the assemble function for the
same reason.

I am adding the `stats_header` attribute to both abctoolbox worker
classes to allow confirmation that it is using he correct stats
for rejection and regression.

The tests of all combinations of regression/rejection methods
are now more thorough.