Fixing bug in `assemble_rejection_workers`.
Within the loop over the lines of the observed simulations file,
where I used the variable `i` to enumerate the line numbers, I also
used `i` within list comprehensions. This was causing all regression
workers to write to the same file, because the index reflected the
last state of the list comprehensions, rather than the line
(simulation) number. This commit corrects this bug.