Adding error message to `ABCToolBoxRejectWorker`.
If the prior file is too large, ABCestimator will seg fault but still
exit with code 0. Thus, the code crashes when an instance of
`ABCToolBoxRejectWorker` goes to post process the posterior file.
This commit adds a check for the posterior files existance during
the `_post_processing` method, and raises an Exception if it does
not exist, and prints the stderr from the ABCestimator subprocess.

This should create clearer error messaging when ABCestimator chokes
on a prior file that is too large.