Updating `workers` module.
Adding `summary_worker` and `rejection_worker` attributes to
`MsBayesWorker` and updating postprocessing method of `MsBayesWorker`
class to run these new workers. This change is intended to allow
EuRejectWorkers to immediately calculate the means and standard
deviations of statistics and/or perform rejection on a prior generated
by the `MsBayesWorker`.

Also, updating `EuRejectWorker` class. Adding the parsing of
the worker's standard error to store summaries of how the
worker ran.

Also, adding the `parse_summary_file` method.

Adding associated tests.