Updating `stats` module.
Fixing error in precondition checks in `get_bin_width` function;
changing `samples < 2` to `len(samples) < 2`.

Also, in `mode_list` function, adding check of bin width returned from
`get_bin_width`. If the returned value is zero, the function tries
setting it to range/10, if that is still zero, it sets the bin width
arbitrarily to 0.001. This was needed to avoid zero division error.