Updating `_run_workers` method of `ABCTeam`.
Adding a `queue_max` option (default 1000). The method now uses this
number to load the work queue in batches of size `queue_max`. This
should prevent the queue from filling, which causes the method to hang
because the number of returning workers in the result queue is less than
expected.