Overhauling `run` method of `Manager`.
Separating the getting and putting of workers onto/off of the queue into
`_get_worker` and `put_worker` methods, respectively. These methods have
checks to try to ensure that managers keep running while jobs are on the
queue, and that they put their finished workers back in the result
queue.