Updating `run_workers` method of `ABCTeam`.
Adding more explicit looping over the number of processors to create,
run, and join `Manager` instances. In some cases, the number of managers
running in parallel is less than the number of processors and jobs in
the queue. This is an attempt to fix this behavior (or at least rule out
potential causes).