Beginning overhaul of `teams` module.
Starting to overhaul the `teams` module on the "teams" branch.  The
original implementation was very memory intensive when running analyses
with tens of thousands of simulations (e.g., storing tens of thousands
of of `RejectionTeam`s in memory for the entire duration of the
analysis). This commit is the first step towards revamping the entire
module to reduce memory consumption. This module will be broken on this
branch for a while.