Updating clean up methods of `TempFileSystem`.
Adding `clear_dir` method, which is now called within `remove_dir`.
This allows a temp dir to be cleared without deleting it. Also, changing
`_remove_file` method to `remove_file`, which now checks if it is
registered, then checks its token, and throws an error if the file
should not be deleted. This logic used to be inside `remove_dir`.