Adding `run_workers` class method to `Manager`.
Adding classmethod `run_workers` to class `Manager`. This is the
primary method by which to use the Manager class to multiprocess
workers. All multiprocessing in this package should now be done
through this method.

In short, the method loads a multiprocessing queue with provided
workers, creates `num_processors` Manager instances to run the
workers in parallel, and then checks all returned workers for
errors, and raises any errors encountered.

Not only is it nice to have one method for carrying out all
multiprocessing, but it also allows all workers of the base
class `Worker` to return and raise their errors to the main
process. This prevents deadlock and hanging when workers raise
exceptions in a subprocess.