Updating `MsBayesConfig` class.
Adding support for parsing and retaining bottleneck magnitude parameter
settings. The distribution on the bottleneck proportion parameters is
stored in the `bottle_proportion` attribute, and whether or not the
descendant populations share a single bottleneck proportion parameter is
stored in the `bottle_proportion_shared` attribute.

Also, the rows of the sample table are now stored (as lists) in the
`sample_table` attribute, and the set of taxon names is stored in
`taxa`.

Also, adding several functions (`_get_gamma_or_uniform_settings`,
`_get_old_settings`, `_get_new_settings`, `get_settings`,
`get_preamble`, `get_sample_table`, `__str__`, and `write`) for
retrieving aspects of the config file settings or re-creating the config
entirely.