Adding spreadsheet functions to `parsing` module.
Adding `spreadsheet_iter` and `get_dict_from_spreadsheets` functions to
the `parsing` module. The former returns an iterator over the lines of
one or more spreadsheets that yields a dict for each line. The latter
returns a dict of all lines of one or more spreadsheets by iterating
over the former and accumulating a dict of lists.

Adding associated tests to `test_parsing.py`.