Updating `dmc.py`.
Removing `InfoLogger` and `argparse`-related functions; these are
now imported from other modules.

Moving most of the within-package importing to within the `main_cli`
function. This is cleaner and allows the logger to be a local variable
within the function.

Also changing the `write_stats_file` attribute of the
observed-stats-generating `MsBayesWorker`s to `False`, because these
files are no longer used.

The script is currently passing tests.