More work on `PlotGrid` and `saturation_plot`.
Adding control of figure size to `PlotGrid` class. Adding check for at
least two subplots in `saturation_plot` function, before setting up a
two-column `PlotGrid`.

Also, adding an `auto_height` attribute to `PlotGrid`, which if
True (default) it will try and calculate a decent figure height
based on the number of rows and columns.

Some adjustments to margin padding in `PlotGrid` as well.