Debugging histogram support in `plotting`.
Adding `Ticks` class for holding ticks and tick labels settings.
Instances of this class can be stored in `ScatterPlot` to allow
ticks and tick labels to be reproduced faithfully when plots are
reset.

Also adding x and y limit attributes to `ScatterPlot` so that the
limits can be reproduced on resets.

Working on debugging `PsiPowerPlotGrid` class.