Updating `DMCSimulationResults`.
Fixing bug in `flat_result_iter` when `include_tau_exclusion_info`
is `True`, but exclusion is not possible for the models. This
resulted in an empty `bf_tau_max` list that `max` was called on,
causing an error. Now if the list is empty, `bf_tau_max` is set
to infinity (i.e., no exclusion is possible).