More work on `DMCSimulationResults.flat_result_iter`.
There were still zero-division issues when comparing models where tau
exclusion was not possible and `include_tau_exclusion_info` was `True`.
Adding checks to see if the prior odds are (almost) zero before
calculating the Bayes factors. If it is, the bf is set to zero or
infinity depending whether the posterior odds are (almost) zero or not.