Updating `PlotGrid` class.
Updating `set_shared_y_limits` method of `PlotGrid` to allow sharing of
y-axis scale `by_row` of the grid of plots. This entailed adding an
optional `plots` argument (which default to all subplots of the
instance) to the `get_widest_y_limits` method.

In order for the `by_row` option of `set_shared_y_limits` to work,
the instance's `share_y` attribute needs to be `False`.