Updating `plotting` module.
Adding code for generating plots of marginal divergence times from the
results in the posterior summary file of a model that samples over
ordered models of divergence.

Adding `ErrorData` class to store the data and perform the plotting.
Adding the `error_data_list` attribute to `ScatterPlot`, along with the
`_plot_error_data` method, so that the `ScatterPlot` class can be used
to generate the divergence-time plot.

Also, adding the `get_marginal_divergence_time_plot` function that wraps
up the creation of the plot via `ScatterPlot`, and sticks the resulting
`ScatterPlot` instance into an instance of `PlotGrid`, and returns the
`PlotGrid`.