More hacking of `plotting` module.
Adding `x_label_size` and `y_label_size` attributes to `ScatterPlot`
class for controling the font size of the x and y labels.

Updating `get_marginal_divergence_time_plot` function to allow control
of x and y label font sizes. Also, creating a `Ticks` instance for the
x-axis. This keeps the x-axis from switching to a serif font when
matplotlib is setup to `usetex`.