Updating `plotting` module.
Overhauling `ErrorData` and `UnorderedDivergenceModelPlotGrid` classes
to allow for plotting of unordered divergence model results.

Also, adding `add_to_ax` method to `TextObj` class. This allows text
annotations on plots in the same coordinates as the data. Adding a
`text_objects_for_ax` attribute to the `ScatterPlot` class, which is
meant to hold `TextObj` instances on which the new `add_to_ax` method is
to be called.

A few other minor tweaks.