Adding new line at end of `test_util_functions.py`.
Adding new line at end of `test_util_functions.py` test script. This
should make no difference, but on some systems (an iPlant Ubuntu 12.04
virtual machine running Python version 2.7.3), this test module is not
loading, and this is the only difference I can see from the other test
scripts.

When `loadTestsFromNames` is called in `pymsbayes.test.__init__.py`,
it results in the error:

    AttributeError: 'module' object has no attribute 'test_util_functions'

`test_util_functions.py` is the only test script without a carriage
return after unittest.main(). Adding this here, but again, I cannot
imagine that this will actually fix the problem. But, it will eliminate
a variable.