Adding `configobj` requirement to `setup.py`.
Adding `configobj` to `install_requires` in `setup.py`. This should fix
the problem discussed in last commit about unittest failing with an
`AttributeError`. Running tests via setuptools was causing the import
error to be "hidden," leading to the misleading `AttributeError`.

This is not surprising, because I forgot to add `configobj` to the
requirements.