Updating test suite.
Adding `test_end_user.py`. This test module contains a suite of tests
for end users to run to ensure the bundled tools are working. The
default testing level is now `FAST`, in which case, the tests in
`test_end_user.py` are the only tests that get loaded and run.

Also, updating the run-level behavior of long-running tests throughout
the test suite to only run when the test level is `EXHAUSTIVE`.