Updating `MSBAYES_SORT_INDEX` and tests.
Updating `MSBAYES_SORT_INDEX` to allow additional indices that have been
added as options to the `dpp-msbayes` package. Updating tests
accordingly.

Also, adding test of new sort index 11 to `test_msbayes_priors.py`,
which will soon be the new default. This test currently fails as
expected. It should not pass until the bundled `dpp-msbayes` executables
are updated to the new versions that offer the new sort-index options.
That is the next step.