Updating package-wide default for sorting index.
Updating `MSBAYES_SORT_INDEX` class so that the default sort index for
the package is 0 (no grouping or sorting) instead of 7. The main CLI
program `dmc.py` still has a default of 7, and needs to be changed to 0.

Updating all tests that required sort index 7 to pass. Explicitly
setting `MSBAYES_SORT_INDEX.set_index(7)` for such tests.