Overhauling tool path management.
Overhauling how tool executables are located. Adding `ToolPathManager`
class to `pymsbayes.utils.__init__`, which handles the locating of
tools, and replaces the old `get_tool_path` function.

Updating the rest of the package to use `ToolPathManager`. This mainly
entailed updating worker and team classes to prevent them from using
`expand_path` method on exe paths.

The new logic for finding a tool `name` is: (1) use `name` if it is a
full path to an exe, (2) check for a bundled exe named `name`, (3) check
the system for an exe `name`. This should allow end users to `python
setup.py install` the package as long as the tools used by the package
are in their PATH. Previously, this would break the package, because the
installed egg did not know where the bundled exectables were located.
This was not a big problem, because `python setup.py develop` worked
fine. However, end users may want to `install`.