Updating `ToolPathManager` class.
Updating `which` method to first check if a name or path was given as
the argument. If it's a path, the method simply checks to see if it's
executable and either returns the path if it is or `None` if it isn't.
If the arg is a name, the system PATH is searched for the path to an
executable of that name.

Also, changing some variable names for clarity.

Adding more tests of `ToolPathManager` to `test_util_functions.py`.