Fixing bug where `MsBayesWorker` would hang.
If an `MsBayesWorker` was launched with `sample_size` zero, it would
hang. Updating `MsBayesWorker.__init__` to raise a `ValueError` if the
`sample_size` is less than 1.

Also updating `dmc.py` script and `ABCTeam` class to avoid ever trying
to create an `MsBayesWorker` with an invalid sample size.  It was
happening in `ABCTeam._prior_summary_worker_iter` if the
`num_standardizing_samples` was greater than `num_prior_samples`, and
the `batch_size` was a multiple of `num_prior_samples`.  Fixing the
logic in this method to avoid the error, and also updating `dmc.py` and
`ABCTeam.__init__` to simply set `num_standardizing_samples` to
`num_prior_samples` if the former is greater than the latter.