Fixing intra-package imports and re-arranging.
The `dmc_plot_results.py` script was not working due to import errors.
The package wide importing of modules, classes, and functions was a mess
and entailed some cross importing between modules.

In this commit, I am cleaning up import statements across much of the
package, and moving some code around to avoid cross importing between
modules. The main issue was that the `parsing` module is imported by
many other modules, and some newer additions to the `parsing` module
required imports from several other modules (`teams`, `plotting`, and
`stats`) that also import from `parsing`. So, I moved the code that
required these modules out of `parsing` and into classmethods of
`workers.PosteriorWorker`, `plotting`, and a new module `sumresults`.
The cleaning and rearranging had rippling affects through a lot of the
package, but all the tests within the test suite are currently passing,
and all the `dmc*` scripts are working.