Adding support for coefficient of variation.
Adding support for the addition of the coefficient of variation of
divergence times ('PRI.cv' column) that was added to ``dpp-msbayes``.
This required changes mostly to the ``parsing`` and ``workers`` modules
so that the new column will be handled the same as the 'PRI.omega'
column.

NOTE, the new versions of dpp-msbayes executables are not being included
in this commit (these will be next), so the changes in this commit will
temporarily 'break' the package.