Updating ABCTeam class.
Updating ABCTeam class to keep better track of the number of samples
that have been processed. The 'num_samples_processed' attribute is now a
dictionary of dictionaries, with the first keys being the observed
index, and the second keys being the model index. The values are the
number of samples processed, which will now correspond to the number of
prior samples processed in each analysis.