Updating config module.
Overhauling config module to more rigorously parse configuration files.
Adding ``SampleTable`` and ``AlignmentConfig`` classes to do all of the
heavy lifting in this regard. Updating ``MsBayesConfig`` class to defer
the parsing of the sample table to these new classes.

Also adding ``SampleTableRowError`` and ``SampleTableError`` classes to
the ``errors`` module. These are used by the new config module classes.

NOTE: All of these changes are completely untested and almost certainly
break the config module as of this commit.

Starting all of this work on the "config" branch.