Adding vetting of configs to dmc.py script.
Adding detailed vetting of configs, both before and after obsSumStats.pl
is run to make sure all of the sample tables are identical.

Changes did not seem to break script (all tests passing), but tests of
new vetting behavior need to be added to make sure differences in sample
tables are caught.