Moving and updating dmc.py code.
Moving the ``main_cli`` function of the dmc.py script to the new
pymsbayes.cli.main_dmc module. The function now takes the argument
``argv = sys.argv``. The main purpose of this move is to improve
testing; the main script function can now be called to check for
specific errors and exceptions. This also keeps more code 'within' the
package, making it more portable and reusable.

Also tweaking the code and error messages within the ``main_cli``
function related to vetting the config files. Adding tests of this new
vetting behavior to ``test_dmc.py``, and all are passing.