Updating dmc_plot_results.py script.
Changing '--iteration-index' option to '--sample-index'
and updating help message to reflect the new behavior of dmc.py
to report file names with the number of prior samples, rather
than the number of batch iterations.

Updating help message and handling of the '--mu' option. The script now
checks the prior config to see if time is in substitutions per site or
not, and the time multiplier is calculated from "--mu" accordingly.